DRL_TurboEngineV2/WorldBuilding/requirements.txt
absl-py==2.2.2
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
etils==1.12.2
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.58.2
fsspec==2025.3.2
glfw==2.9.0
gymnasium==1.1.1
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
mujoco==3.3.2
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
zipp==3.21.0
